requirements.txt
Flask==2.0.1
google-cloud-storage==1.42.2
firebase_admin
gunicorn==20.0.4